restructure (rename folder src to app), adding online_presence, adding to README.md,
