+++ .vscode/settings.json
@@ -2,6 +2,6 @@
     "cornflakes.linter.executablePath": "venv/bin/flake8",
     "cornflakes.linter.run": "onType",
     "python.analysis.extraPaths": [
-        "./src"
+        "./app"
     ]
 }
