+++ README.md
@@ -9,7 +9,7 @@ Written for python 3.10
 # Licencing and acknowledgements
 The project is published under the [MIT Licence](LICENSE.md).
 
-The data bundled in [src/data/helmgast/](src/data/helmgast/) is created and owned by Helmgast AB.</br>
+The data bundled in [app/data/helmgast/](app/data/helmgast/) is created and owned by Helmgast AB.</br>
 Permission has been granted by representatives of Helmgast AB to extract the hit table & damage table data from the books and bundle them in this project.
 
 # Usage
@@ -65,7 +65,7 @@ python3 -m pip install -U py-cord --pre
 ```
 
 # CONFIGURATION!
-write a .env file based from src/.env.example
+write a .env file based from app/.env.example
 generate an api token from [discord's developer portal](https://discord.com/developers/).
 Join bot to your discord server from the developer portal.
 
@@ -77,7 +77,7 @@ https://discord.com/developers/
 ```
 cd ~/git/eon-bot
 source venv/bin/activate
-python3 src/main.py
+python3 app/main.py
 ```
 
 ## Run Docker container
@@ -95,7 +95,7 @@ chmod +x start_dockerized.sh
 docker build . -t eon-bot:1.3.0
 # Run
 docker run -d \
- --mount type=bind,source=$PWD/src/.env,target=/app/src/.env,readonly \
+ --mount type=bind,source=$PWD/app/.env,target=/eon-bot/app/.env,readonly \
  --restart unless-stopped \
  --name eon-bot \
  eon-bot:1.3.0
@@ -106,7 +106,11 @@ docker run -d \
 Slash commands! You may use discord's slash commands in order to use the bot.
 
 ## Syntax
-### /roll
+### /dm - Direct message
+eon-bot will send a direct message to the user.
+Useful for secret dice rolls or lookups.
+
+### /roll - Scalable dice
 Format has to be in NtN+N, NtN, NdN+N or NdN.
 
 Can also be written with capital letters: ie NT6+N
@@ -122,7 +126,7 @@ Output example:
 
 ```
 
-### /ob
+### /ob or /inf - Infinite dice
 Format has to be in Nt6+N, Nt6, Nd6+N or Nd6.
 
 if anything else than a 6 is supplied, it will be ignored. T6/D6 dice are hardcoded for the ob dice.
@@ -137,13 +141,16 @@ Example
 /ob 4d6+3
 
 # This example will still use a six sided die.
-/ob 4t8 
+/ob 4t8 +3
 
 Output example:
-
+Rolled : 4T6+3
+No. Sixes.... : 2
+Rolls............ : 2, 1, 3, 3, 4, 4+ 3
+Total............ : 20
 ```
 
-### /fight
+### /fight - Fight assistant
 Will roll ob +t100 dice for use in a fight.
 User runs ```/fight <diceroll of weapon>```
