+++ start_dockerized.sh
@@ -7,6 +7,6 @@ IMAGEVER="1.3.0"
 docker build . -t ${CONTAINERNAME}:${IMAGEVER}
 # Run Image
 docker run -d \
- --mount type=bind,source=$PWD/src/.env,target=/app/src/.env,readonly \
+ --mount type=bind,source=$PWD/app/.env,target=/eon-bot/app/.env,readonly \
  --restart unless-stopped \
  ${CONTAINERNAME}:${IMAGEVER}
